refactor(index.js): 가독성을 높이기 위해 함수 모듈화
- BaseballGame.js: index.js에서 BaseballGame 분리
- EventManager.js: 이벤트 리스너와 핸들러 관리
- helpers.js: querySelector, addEventListener wrapper 함수
