+++ src/helpers.js
@@ -1,0 +1,9 @@
+// querySelector wrapper
+export function $(selector) {
+  return document.querySelector(selector);
+}
+
+// addEventListener wrapper
+export function $on(target, eventType, callback, capture) {
+  target.addEventListener(eventType, callback, !!capture);
+}
